Move, merge & rename networking implementations in shared project
For efficiency, we keep things like the Send() implementation implemented directly in the implementing class.
They tend to handle these differently, so we shouldn't try to one-size-fits-all them, even though most are similar.

We also hide away direct access to SocketManager's (formerly CIOCPort) members, so we can more consistently verify them.

To avoid any awkward deadlocks (particularly with AI connections in Ebenezer), we ensure listbox updates are done only within
the window's main thread. We queue them and post a message to trigger their update when it gets a chance.
Otherwise it will use SendMessage(), blocking until the main thread processes, which can be particularly bad if the main thread
is doing things with locks already, (e.g. already trying to connect to AI).

+++ Server/Ebenezer/MagicProcess.h
@@ -28,9 +28,9 @@ class CMagicProcess
 {
 public:
 	int16_t GetWeatherDamage(int16_t damage, int16_t attribute);
-	void SendType4BuffRemove(int16_t tid, uint8_t buff);
-	void Type3Cancel(int magicid, int16_t tid);
-	void Type4Cancel(int magicid, int16_t  tid);
+	void SendType4BuffRemove(int tid, uint8_t buff);
+	void Type3Cancel(int magicid, int tid);
+	void Type4Cancel(int magicid, int tid);
 	bool UserRegionCheck(int sid, int tid, int magicid, int radius, int16_t mousex = 0, int16_t mousez = 0) const;
 	int16_t GetMagicDamage(int sid, int tid, int total_hit, int attribute);
 	void ExecuteType10(int magicid);
